% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_barro_lee.R
\name{get_barro_lee}
\alias{get_barro_lee}
\title{Download Barro-Lee educational data}
\usage{
get_barro_lee(download_data, countries_considered, first_year, last_year)
}
\arguments{
\item{download_data}{If TRUE, data will be downloaded from the internet}

\item{countries_considered}{Vector of iso3c codes of the countries for
which data should be assembled.}

\item{first_year}{First year for which data is to be collected (numeric).}

\item{last_year}{Last year for which data is to be collected (numeric).}
}
\description{
Get educational data from Barro-Lee dataset
}
\seealso{
Other download_helpers: \code{\link{get_ameco}},
  \code{\link{get_chinn_ito}},
  \code{\link{get_complexity}}, \code{\link{get_kof}},
  \code{\link{get_oecd_data}}, \code{\link{get_solt}},
  \code{\link{get_worldbank}}
}
\concept{download_helpers}
