% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chinn_ito.R
\name{get_chinn_ito}
\alias{get_chinn_ito}
\title{Get data on the Chinn-Ito index}
\usage{
get_chinn_ito(download_data, countries_considered, first_year, last_year)
}
\arguments{
\item{download_data}{If TRUE, data will be downloaded from the internet}

\item{countries_considered}{Vector of iso3c codes of the countries for
which data should be assembled.}

\item{first_year}{First year for which data is to be collected (numeric).}

\item{last_year}{Last year for which data is to be collected (numeric).}
}
\description{
Downloads data directly from the Chinn-Ito website
}
\seealso{
Other download_helpers: \code{\link{get_ameco}},
  \code{\link{get_barro_lee}},
  \code{\link{get_complexity}}, \code{\link{get_kof}},
  \code{\link{get_oecd_data}}, \code{\link{get_solt}},
  \code{\link{get_worldbank}}
}
\concept{download_helpers}
