% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R, R/macro_data.R
\name{get_countries}
\alias{get_countries}
\title{Get countries}
\usage{
get_countries(countries_to_get)

get_countries(countries_to_get)
}
\arguments{
\item{countries_to_get}{A code for a pre-specified country list}

\item{countries_to_get}{A code for a pre-specified country list}
}
\value{
A vector with country codes

A vector with country codes
}
\description{
Returns a pre-specified country list

Returns a pre-specified country list
}
\seealso{
Other update_dataset_helpers: \code{\link{test_uniqueness}},
  \code{\link{unfactor}}
}
\concept{update_dataset_helpers}
