% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eurostat.R
\name{get_eurostat_bond_data}
\alias{get_eurostat_bond_data}
\title{Get data on bond yields from Eurostat}
\usage{
get_eurostat_bond_data(download_data, countries_considered, first_year,
  last_year)
}
\description{
Gets data on bond yields from Eurostat from dataset 'irt_lt_mcby_a'.
}
\details{
@param download_data If TRUE, data will be downloaded from the internet
 @param countries_considered Vector of iso3c codes of the countries for
  which data should be assembled.
 @param first_year First year for which data is to be collected (numeric).
 @param last_year Last year for which data is to be collected (numeric).
 @family download_helpers
}
