% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kof.R
\name{get_kof}
\alias{get_kof}
\title{Get KOF data}
\usage{
get_kof(download_data, countries_considered, first_year, last_year)
}
\arguments{
\item{download_data}{If TRUE, data will be downloaded from the internet}

\item{countries_considered}{Vector of iso3c codes of the countries for
which data should be assembled.}

\item{first_year}{First year for which data is to be collected (numeric).}

\item{last_year}{Last year for which data is to be collected (numeric).}
}
\description{
Download the KOF globalization index
}
\details{
Downloads and assembles data from the KOF globalization index project.
 Since the KOF institute does not provide for an API the data is downloaded
 directly from their homepage. This means that whenever a new version of the
 KOF is availbale, this code must be updated. Latest update: May 2019.
}
\seealso{
Other download_helpers: \code{\link{get_ameco}},
  \code{\link{get_barro_lee}}, \code{\link{get_chinn_ito}},
  \code{\link{get_complexity}},
  \code{\link{get_oecd_data}}, \code{\link{get_solt}},
  \code{\link{get_worldbank}}
}
\concept{download_helpers}
