% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lmf.R
\name{get_lmf}
\alias{get_lmf}
\title{Get data from LMF}
\usage{
get_lmf(download_data, countries_considered, first_year, last_year)
}
\arguments{
\item{download_data}{If TRUE, data will be downloaded from the internet}

\item{countries_considered}{Vector of iso3c codes of the countries for
which data should be assembled.}

\item{first_year}{First year for which data is to be collected (numeric).}

\item{last_year}{Last year for which data is to be collected (numeric).}
}
\description{
Assembles relevant indicators from LMF database
}
\details{
Assembles relevant indicators from the Lane-Milesi-Ferreti database. Because
 data is published only in Excel, the dataset as such must be downloaded
 manually and placed in the data-raw folder. It then can be processed using
 this function. This is done to avoid any functions that rely on JAVA, since
 this often becomes problematic in practice.
}
\concept{download_helpers
Transforms cols with characters into doubles explicitly

Conducts the following cleaning procedures:
 1. NA values remain NA.
 2. If the string contains the substring 'NA', set to NA
 3. Remove points from strings
 4. Convert to double}
