% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solt.R
\name{get_solt}
\alias{get_solt}
\title{Get data from SWIID 8 database}
\usage{
get_solt(download_data, countries_considered, first_year, last_year)
}
\arguments{
\item{download_data}{If TRUE, data will be downloaded from the internet}

\item{countries_considered}{Vector of iso3c codes of the countries for
which data should be assembled.}

\item{first_year}{First year for which data is to be collected (numeric).}

\item{last_year}{Last year for which data is to be collected (numeric).}
}
\description{
Downloads data from swiid 8.0 via Harvard Dataverse
}
\details{
Gets pre and post tax Ginis from the standardized world income inequality
 database provided by Solt.
}
\seealso{
Other download_helpers: \code{\link{get_ameco}},
  \code{\link{get_barro_lee}}, \code{\link{get_chinn_ito}},
  \code{\link{get_complexity}}, \code{\link{get_kof}},
  \code{\link{get_oecd_data}}, \code{\link{get_worldbank}}
}
\concept{download_helpers}
