% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macro_data}
\alias{macro_data}
\title{Macroeconomic variables.}
\format{A data on macroeconomic indicators with X rows and Y variables:
\describe{
  \item{iso3c}{country, iso3c code}
  \item{year}{year of observation}
  \item{res_rents}{weight of the diamond, in carats}
  \item{gini_post_tax}{gini from SWIID, post tax (disp inc)}
  \item{gini_pre_tax}{gini from SWIID, pre tax (market)}
  ...
}}
\source{
SWIID 8.0, World Bank WDI, ...
}
\usage{
macro_data
}
\description{
A dataset on macroeconomic indicators.
}
\keyword{datasets}
