% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R, R/macro_data.R
\name{unfactor}
\alias{unfactor}
\title{Unfactor a factor}
\usage{
unfactor(x)

unfactor(x)
}
\arguments{
\item{x}{An input that is potentially a factor}

\item{x}{An input that is potentially a factor}
}
\value{
x as an integer

x as an integer
}
\description{
Transforms a factor into an integer

Transforms a factor into an integer
}
\details{
Transforms a factor into an integer by first transforming it into character

Transforms a factor into an integer by first transforming it into character
}
\seealso{
Other update_dataset_helpers: \code{\link{get_countries}},
  \code{\link{test_uniqueness}}
}
\concept{update_dataset_helpers}
