% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_macro_data.R
\name{update_macro_data}
\alias{update_macro_data}
\title{Updata data}
\usage{
update_macro_data(update_full = FALSE, vars = "all")
}
\arguments{
\item{vars}{Specifies the variables to be updated as strings. Default is all.}

\item{updata_full}{Should all data be downloaded anew (if true), or should
previously downloaded data be used (if false)?}
}
\value{
Returns the updated data set.
TODO: Maybe only information?
}
\description{
\code{update_data} function updates the data.
}
\details{
\code{update_data} function updates the data set. To function properly it
 requires access to the internet. The data can also be updated selectively.
}
